pare down functions.js; update config checks and README
